requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
langchain==0.2.0
langchain-groq==0.0.3
langgraph==0.0.41